Remove cache contingency on Zephyr version
There's no way for us to check ex ante which Zephyr version is pulled in
by west. We could follow the upstream approach and export it from the
container or nix environment. But this seems silly since the toolchain
generally is not specific to a specific Zephyr version.

In either case, the downside seems small since it would only be
beneficial if we'd upgrade the container / nix environment (which for
now is hard-coded) and _then_ revert the upgrade.